Reminders and Pixy Requirements

+++ src/app/cal-utils/client/client-input/client-input.component.ts
@@ -9,7 +9,8 @@ import {Client} from '../shared/client.model';
 import {ClientService} from '../shared/client.service';
 import {ClientAppointmentsService} from '../shared/client-appointments.service';
 
-
+import { AngularFireDatabase, AngularFireList } from 'angularfire2/database';
+import { Subject } from 'rxjs/Subject';
 
 @Component({
   selector: 'app-client-input',
@@ -21,27 +22,37 @@ export class ClientInputComponent implements OnInit {
   customerForm: FormGroup;
   customer: Client= new Client();
   emailMessage: string;
+  firstNameMessage: string;
   private sub: Subscription;
 
 
   mask: any[] = ['+', '1', /[1-9]/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/];
 
 
-  private validationMessages = {
+  private emailvalidationMessages = {
     required: 'Please enter your email address.',
     pattern: 'Please enter a valid email address.'
 };
 
+private firstNamevalidationMessages = {
+  required: 'Please enter Clients Name.',
+  minlength: 'Name must be longer than 3 characters.'
+};
+
 pageTitle: string = 'Product Edit';
 
-
+get landline(): FormArray {
+  return <FormArray>this.customerForm.get('landline');
+}
+refresh: Subject<any> = new Subject();
 
 constructor(private fb: FormBuilder, 
             private route: ActivatedRoute,
             private router: Router, 
             public clientService: ClientService, 
             private clientAppService: ClientAppointmentsService, 
-            private toastr: ToastrService) { }
+            private toastr: ToastrService,
+            private firebase : AngularFireDatabase) { }
 
 
   ngOnInit() {
@@ -54,7 +65,8 @@ constructor(private fb: FormBuilder,
           lastName: ['', [Validators.maxLength(50)]],
           email: ['', [Validators.pattern('[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+')]],
           phone: ['', [Validators.pattern('[0-9+]{12}')]],
-          landline: ['', [Validators.pattern('[0-9+]{12}')]],
+          //landline: ['', [Validators.pattern('[0-9+]{12}')]],
+          landline: this.fb.array([]),
           age: '',
           $key: null,
           gender: '',
@@ -64,7 +76,10 @@ constructor(private fb: FormBuilder,
           postal: '',
           profession: '',
           notes: '',
-          lead: ''
+          lead: '',
+          reminder: '',
+          status: 'Not Started',
+          createDate: new Date()
       });
 
       // Read the product Id from the route parameter
@@ -75,19 +90,43 @@ constructor(private fb: FormBuilder,
         }
     );
 
-      this.customerForm.valueChanges.subscribe(value=> 
-        this.clientService.selectedClient = value);
+      // this.customerForm.valueChanges.subscribe(value=> 
+      //   this.clientService.selectedClient = value);
 
         const emailControl = this.customerForm.get('email');
-        emailControl.valueChanges.subscribe(value =>
-          this.setMessage(emailControl));
+        emailControl.valueChanges.debounceTime(1000).subscribe(value =>
+          this.setMessage(emailControl, "email"));
+
+          const firstNameControl = this.customerForm.get('firstName');
+          firstNameControl.valueChanges.debounceTime(1000).subscribe(value =>
+          this.setMessage(firstNameControl, "firstName"));
   }
 
-  setMessage(c: AbstractControl): void {
-    this.emailMessage = '';
-    if ((c.touched || c.dirty) && c.errors) {
-        this.emailMessage = Object.keys(c.errors).map(key =>
-            this.validationMessages[key]).join(' ');
+  addAlternatePhone(): void {
+    this.landline.push(new FormControl());
+  }
+
+  deleteAlternatePhone(index: number): void {
+    this.landline.removeAt(index);
+    // The line below is required in Angular 4 to fix a bug with `removeAt` that was fixed in Angular 5.
+    //this.productForm.setControl('tags', this.fb.array(this.tags.value || []));
+  }
+
+  setMessage(c: AbstractControl, control: string): void {
+
+    if(control == "email"){
+      this.emailMessage = '';
+      if ((c.touched || c.dirty) && c.errors) {
+          this.emailMessage = Object.keys(c.errors).map(key =>
+              this.emailvalidationMessages[key]).join(' ');
+      }
+    }
+    else if(control == "firstName"){
+      this.firstNameMessage = '';
+      if ((c.touched || c.dirty) && c.errors) {
+        this.firstNameMessage = Object.keys(c.errors).map(key =>
+            this.firstNamevalidationMessages[key]).join(' ');
+    }
     }
 }
 
@@ -99,7 +138,7 @@ constructor(private fb: FormBuilder,
         firstName: ' ',
         lastName: ' ',
         phone: ' ',
-        landline: ' ',
+        landline: [' '],
         email: ' ',
         age: ' ',
         gender: ' ',
@@ -109,7 +148,10 @@ constructor(private fb: FormBuilder,
         postal: '',
         lead: '',
         profession:'',
-        notes:''
+        notes:'',
+        reminder:'',
+        status: 'Not Started',
+        createDate: null
 
     }
 
@@ -118,7 +160,7 @@ constructor(private fb: FormBuilder,
       lastName: '',
       email: '',
       phone: '',
-      landline: '',
+      //landline: '',
       age: '',
       $key: null,
       gender: '',
@@ -128,37 +170,51 @@ constructor(private fb: FormBuilder,
       postal: '',
       lead: '',
       profession: '',
-      notes: ''
+      notes: '',
+      reminder: '',
+      status: 'Not Started',
+      createDate: ''
     });  
-
+    this.customerForm.setControl('landline', this.fb.array('' || []));
   }
 
   save() {
+
+    this.clientService.selectedClient = this.customerForm.value;
     console.log(this.customerForm);
     console.log('Saved: ' + JSON.stringify(this.customerForm.value));
 
     if(this.customerForm.value.$key == null || this.customerForm.value.$key == ''){
-      var k = this.clientService.insertClient(this.customerForm.value);
+      var k = this.clientService.insertClient(this.clientService.selectedClient);
       this.clientAppService.insertNewClient(k);
+      this.router.navigate(['/welcome']);
     }
     else{
       console.log(this.clientService.selectedClient);
       this.clientService.updateClient(this.clientService.selectedClient);
     }
      this.resetForm(this.customerForm);
-     this.router.navigate(['/welcome']);
+     //this.router.navigate(['/welcome']);
      this.toastr.success('Submitted Succcessfully', 'Client');
   }
 
   onDelete(key: string){
-    if(confirm('Are you sure you want to remove this record?')== true){
-      this.clientService.deleteClient(key);
-      this.toastr.warning('Deleted Successfully', 'Client Register');
+    if ( this.clientService.selectedClient.$key === null) {
+      this.router.navigate(['/welcome']);
+    } 
+    else {
+      if(confirm('Are you sure you want to remove this record?')== true){
+        this.clientService.deleteClient(key);
+        this.toastr.warning('Deleted Successfully', 'Client Register');
+        this.router.navigate(['/welcome']);
+      }
     }
+
+   
   }
 
 
-  
+  //this is called from routing url, if 0 add  client if not edit client
   getClient(key: string){
     console.log(key);
 
@@ -188,33 +244,86 @@ constructor(private fb: FormBuilder,
     onEdit(client: Client){
       if (this.customerForm) {
             this.customerForm.reset();
+            
         }
-   
+        //this.resetForm(this.customerForm);
+       // console.log(client);
       this.clientService.selectedClient = client;
+      var stArray;
+     
+     
+      if(client.reminder == undefined){
+        client.reminder = '';
+      }
+      
+      if(client.createDate == undefined){
+        client.createDate = '';
+      }
 
+      if(client.status == undefined){
+        client.status = '';
+      }
+      
+       
+     
       if ( this.clientService.selectedClient.$key === null) {
             this.pageTitle = 'Add Client';
         } else {
             this.pageTitle = `Edit Client: ${this.clientService.selectedClient.firstName} ${this.clientService.selectedClient.lastName}`;
         }
+            
+            console.log(client);
+            console.log(this.clientService.selectedClient);
+            console.log(this.clientService.selectedClient.reminder);
+       
+            this.firebase.database.ref('clients/'+this.clientService.selectedClient.$key+'/landline').once('value', function(snapshot) {
+              console.log(snapshot.val());
+              stArray= snapshot.val();
+              // ["Bill Gates", "Larry Page", "James Tamplin"] from {0:Bill, 1:Larry}
+            }).then(res => {
+              this.customerForm.setControl('landline', this.fb.array(stArray || []));
+             
+          });
+           
+         
+
       this.customerForm.patchValue({
         firstName: client.firstName,
-      lastName: client.lastName,
-      email: client.email,
-      phone: client.phone,
-      landline: client.landline,
-      age: client.age,
-      $key: client.$key,
-      gender: client.gender,
-      street1: client.street1,
-      city: client.city,
-      province: client.province,
-      postal: client.postal,
-      lead: client.lead,
-      profession: client.profession,
-      notes: client.notes
+        lastName: client.lastName,
+        email: client.email,
+        phone: client.phone,
+        //landline: client.landline,
+        age: client.age,
+        $key: client.$key,
+        gender: client.gender,
+        street1: client.street1,
+        city: client.city,
+        province: client.province,
+        postal: client.postal,
+        lead: client.lead,
+        profession: client.profession,
+        notes: client.notes,
+        reminder: client.reminder,
+        status: client.status,
+        createDate: client.createDate
 
       });  
+    
+    }
+
+    clearReminder(){
+      //this.clientService.selectedClient.reminder = "";
+      this.customerForm.patchValue({
+        reminder: '',
+      });
+      
+
+    }
+
+    setReminderToday(){
+      this.customerForm.patchValue({
+        reminder: new Date(),
+      });
     }
 
 }
